backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
pymysql==1.1.0
cryptography==41.0.7

# Data validation and serialization
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Utilities
python-ulid==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy<2.0.0
shapely==2.0.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
ruff==0.1.6
pre-commit==3.5.0

# Type checking
mypy==1.7.1
types-requests==2.31.0.10